feat: Added a analytic manager class
This class will orchestrate the execution order of the data extraction, and data transformation

Also the logic is changed. The api rest manager will not be in charge of setting the cluster coordinates, the extract will do it. However the same logic is maintained (either consulting the api rest or from local), if the csv file does not exist, the json will be processed to obtain it, otherwise it will be loaded directly from there.